Reorganize database funcs
